+++ src/View/Helper/WidgetHelper.php
@@ -54,10 +54,11 @@ class WidgetHelper extends Helper
         //  and the arguments are the array value
         foreach ($widgets as $name => $args) {
             if (is_int($name) && is_string($args)) {
-                $widgetsCopy[$args] = [];
-            } else {
-                $widgetsCopy[$name] = $args;
+                $name = $args;
+                $args = [];
             }
+
+            $widgetsCopy[$name] = $args;
         }
 
         return $widgetsCopy;
@@ -65,20 +66,18 @@ class WidgetHelper extends Helper
 
     /**
      * Renders all widgets
-     * @return string|null Html code
+     * @return string|void Html code
      * @uses _getAll()
      * @uses widget()
      */
     public function all()
     {
-        $widgets = $this->_getAll();
-
-        if (empty($widgets)) {
-            return null;
+        foreach ($this->_getAll() as $name => $args) {
+            $widgets[$name] = $this->widget($name, $args);
         }
 
-        foreach ($widgets as $name => $args) {
-            $widgets[$name] = $this->widget($name, $args);
+        if (empty($widgets)) {
+            return;
         }
 
         return trim(implode(PHP_EOL, $widgets));