+++ src/View/View/AdminView.php
@@ -30,6 +30,12 @@ use MeCms\View\View\BaseView;
 class AdminView extends BaseView
 {
     /**
+     * The name of the layout file to render the template inside of
+     * @var string
+     */
+    public $layout = 'MeCms.admin';
+
+    /**
      * Initialization hook method
      * @return void
      * @see http://api.cakephp.org/3.3/class-Cake.View.View.html#_initialize
@@ -64,21 +70,5 @@ class AdminView extends BaseView
         ]);
 
         return parent::render($view, $layout);
-    }
-
-    /**
-     * Renders a layout. Returns output from _render(). Returns false on error.
-     *  Several variables are created for use in layout
-     * @param string $content Content to render in a view, wrapped by the
-     *  surrounding layout
-     * @param string|null $layout Layout name
-     * @return mixed Rendered output, or false on error
-     * @see http://api.cakephp.org/3.3/class-Cake.View.View.html#_renderLayout
-     */
-    public function renderLayout($content, $layout = null)
-    {
-        $layout = 'MeCms.admin';
-
-        return parent::renderLayout($content, $layout);
     }
 }